Refactoring of ToolbarCreator
